# perf(*) remove unnecessary timer since no yield needed