--- a/kong/plugins/prometheus/handler.lua
+++ b/kong/plugins/prometheus/handler.lua
@@ -9,15 +9,6 @@ local timer_at = ngx.timer.at
 prometheus.init()
 
 
-local function log(premature, message)
-  if premature then
-    return
-  end
-
-  prometheus.log(message)
-end
-
-
 local PrometheusHandler = {
   PRIORITY = 13,
   VERSION  = "0.4.1",
@@ -26,10 +17,7 @@ local PrometheusHandler = {
 
 function PrometheusHandler:log(_)
   local message = basic_serializer.serialize(ngx)
-  local ok, err = timer_at(0, log, message)
-  if not ok then
-    kong.log.err("failed to create timer: ", err)
-  end
+  prometheus.log(message)
 end